+++ Sources/VCore/Services and Managers/Presentation Host/Presentation Mode/PresentationHostPresentationModeEnvironmentValue.swift.swift
@@ -8,10 +8,6 @@
 import SwiftUI
 
 // MARK: - Presentation Host Presentation Mode Environment Value
-@available(macOS, unavailable)
-@available(tvOS, unavailable)
-@available(watchOS, unavailable)
-@available(visionOS, unavailable)
 extension EnvironmentValues {
     /// Presentation Host's presentation mode of the `View` associated with the environment.
     @EnvironmentValueGeneration public var presentationHostPresentationMode: PresentationHostPresentationMode?