+++ Sources/VCore/Services and Managers/Presentation Host/PresentationHostDataSourceCache.swift
@@ -8,14 +8,15 @@
 import SwiftUI
 
 // MARK: - Presentation Host Data Source Cache
-/// Caches data for Presentation Host, for when it may become `nil`.
+/// Caches data for Presentation Host, for when presenting model may become `nil`.
 ///
 /// Can be used to ensure that last content is visible instead of blank screen, when data is set to `nil` and modal is being dismissed.
 ///
-/// For additional info, refer to `View.presentationHost(id:uiModel:isPresented:onPresent:onDismiss:content:)`.
+/// For additional info, refer to `View.presentationHost(...)`.
 ///
 ///     extension View {
 ///         func someModal<T>(
+///             layerID: String? = nil,
 ///             id: String,
 ///             isPresented: Binding<Bool>,
 ///             presenting data: T?,
@@ -25,6 +26,7 @@ import SwiftUI
 ///
 ///             return self
 ///                 .presentationHost(
+///                     layerID: layerID,
 ///                     id: id,
 ///                     isPresented: isPresented,
 ///                     presenting: data,
@@ -39,10 +41,6 @@ import SwiftUI
 ///         }
 ///     }
 ///
-@available(macOS, unavailable)
-@available(tvOS, unavailable)
-@available(watchOS, unavailable)
-@available(visionOS, unavailable)
 public final class PresentationHostDataSourceCache {
     // MARK: Properties
     /// Shared instance of `PresentationHostDataSourceCache`.